server/requirements.txt
aniso8601==4.0.1
Click==7.0
Flask==1.0.2
Flask-PyMongo==2.1.0
Flask-RESTful==0.3.6
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.0
pymongo==3.7.2
python-dotenv==0.9.1
pytz==2018.6
six==1.11.0
Werkzeug==0.14.1
